\name{Rperio-package}
\alias{Rperio-package}
\alias{Rperio}
\docType{package}
\title{
Tools for Case Definitions and prevalences in Population-Based Surveillance of Periodontitis.}
\description{
A collection of tools for Case Definitions and prevalences in 
Population-Based Surveillance of Periodontitis. 
  The functionality is experimental and functions are likely to changein the future. 
  This is based on the fact that periodontal case definitions is a rapidly changing
  ares. The functions hereby are based on the  CDC Working Group (2010), as well
  as the 5thEuropean workshop in Periodontology case definitions.
  However, if you come acroos ideas interesting we would be very interested in a discussion
  of this proposal. Contributions are more than welcome!
}
\details{
\tabular{ll}{
Package: \tab Rperio\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-01-03\cr
License: \tab What license is it under?\cr
}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
Edwin M. Kagereki

Maintainer: Edwin M. Kagereki <kabethte@gmail.com>

}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ Periodontal data management }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
